team hierarchies

+++ src/app/api/staff/delete/[id]/route.js
@@ -1,20 +1,30 @@
-import { NextRequest, NextResponse } from "next/server";
+import { NextResponse } from "next/server";
 import User from "@/models/Users";
 import connect from "@/dbConfig/dbConfig";
+import { deletePermitedRoles } from "../../permissions";
 
 connect();
 
 export async function DELETE(request, { params }) {
-    try {
-        const userid = params.id;
-        if (!userid) {
-            return NextResponse.json({ error: "User ID is required" }, { status: 400 });
-        }
-
-        await User.deleteOne({ _id: userid });
-
-        return NextResponse.json({ message: "User deleted successfully" });
-    } catch (error) {
-        return NextResponse.json({ error: error.message }, { status: 400 });
+  try {
+    const userid = params.id;
+    if (!userid) {
+      return NextResponse.json(
+        { error: "User ID is required" },
+        { status: 400 }
+      );
     }
+
+    if (deletePermitedRoles.includes(Role))
+      return NextResponse.json(
+        { error: "You don't have permissions to add staff" },
+        { status: 401 }
+      );
+
+    await User.deleteOne({ _id: userid });
+
+    return NextResponse.json({ message: "User deleted successfully" });
+  } catch (error) {
+    return NextResponse.json({ error: error.message }, { status: 400 });
+  }
 }
